pyproject.toml
[project]
name = "dietlog-api"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
  "anthropic>=0.45.2",
  "fastapi[standard]>=0.115.8",
  "httpx>=0.28.1",
  "python-dotenv>=1.0.1",
]

[tool.pyright]
reportMissingTypeStubs = true
allowedUntypedLibraries = []
typeCheckingMode = "recommended"

[tool.ruff]
line-length = 120
indent-width = 4
target-version = "py312"

[tool.ruff.lint]
select = ["ALL"]
ignore = [
  "COM812", # https://docs.astral.sh/ruff/rules/missing-trailing-comma/
  "D203",
  "D213",
  "E501",
]
